Extract reference image queries into a new file
